Refactor Profiler -> SamplingProfiler
