+++ packages/core/profiler/src/index.js
@@ -1,2 +1,3 @@
-export {default as Profiler} from './Profiler';
+// @flow
+export {default as SamplingProfiler} from './SamplingProfiler';
 export {default as Trace} from './Trace';
